restructure : MetricsProvider.jsx moved to contexts and NavMenu.jsx moved to navMenu folder folder
