+++ src/components/ownComponents/metricsPage/StatisticOne/SelectFeelingsFamily.jsx
@@ -11,7 +11,7 @@ import MessageMetricsM1 from "./MessageMetricsM1";
 
 // context providers:
 import { useEmotionsContext } from "@/utils/contexts/EmotionsProvider";
-import { useMetricsContext } from "@/utils/MetricsProvider";
+import { useMetricsContext } from "../../../../utils/contexts/MetricsProvider";
 
 // ------------------------------- SelectFeelingsFamily Component -------------------------------
 // this component allows the user to select a feelings family and loads the corresponding statisitcs
