+++ src/components/ownComponents/metricsPage/StatisticTwo/SelectTag.jsx
@@ -1,5 +1,5 @@
 import { CardHeader } from "@/components/ui/card";
-import { useMetricsContext } from "@/utils/MetricsProvider";
+import { useMetricsContext } from "../../../../utils/contexts/MetricsProvider";
 import { useTagContext } from "@/utils/contexts/TagProvider";
 import MessageMetricsM2 from "./MessageMetricsM2";
 import UserFeedbackText from "@/components/typo/UserFeedbackText";
